+++ packages/terra-date-picker/src/DateInput.jsx
@@ -170,6 +170,8 @@ class DatePickerInput extends React.Component {
           placeholder={placeholder}
           onFocus={onFocus}
           onBlur={onBlur}
+          id="Date-Input"
+          aria-label={this.context.intl.formatMessage({ id: 'Terra.dateInput.date' })}
         />
         <Button
           className={styles.button}