+++ packages/terra-date-time-picker/src/terra-dev-site/test/date-time-picker/twelve-hour/DateTimePickerDisabled12HourMobile.test.jsx
@@ -1,12 +1,12 @@
 import React from 'react';
 import classNames from 'classnames/bind';
-import DateTimePickerExampleTemplate from '../../common/DateTimePickerExampleTemplate';
-import DateTimeUtil from '../../../DateTimeUtils';
+import DateTimePickerExampleTemplate from '../../../common/DateTimePickerExampleTemplate';
+import DateTimeUtil from '../../../../DateTimeUtils';
 import styles from './DateTimePicker12HourMobile.test.module.scss';
 
 const cx = classNames.bind(styles);
 
-export default class DateTimePicker12HourMobile extends React.Component {
+export default class DateTimePicker12HourDisabledMobile extends React.Component {
   constructor(props) {
     super(props);
     this.state = { datetime: '2019-04-10T15:23:00-05:00' };
@@ -37,22 +37,6 @@ export default class DateTimePicker12HourMobile extends React.Component {
   render() {
     return (
       <div className={cx('content-wrapper')}>
-        <h3>Empty Date-Time-Picker</h3>
-        <DateTimePickerExampleTemplate
-          id="datetimeEmpty"
-          timeVariant={DateTimeUtil.FORMAT_12_HOUR}
-        />
-
-        <h3 id="titleWithDateTimeValue">
-          Date-Time-Picker with value provided - Current Value:&nbsp;
-          {this.state.datetime}
-        </h3>
-        <DateTimePickerExampleTemplate
-          id="datetimeValueProvided"
-          value={this.state.datetime}
-          timeVariant={DateTimeUtil.FORMAT_12_HOUR}
-        />
-
         <h3 id="titleWithDateTimeValue">Disabled Date-Time-Picker</h3>
         <DateTimePickerExampleTemplate
           id="disabledDatetimeValueProvided"