+++ packages/terra-date-picker/CHANGELOG.md
@@ -4,6 +4,9 @@ Changelog
 Unreleased
 ----------
 ### Added
+* Aria-Label and ID added to input to fix accesssibility issue of Date-Time-Picker.
+
+### Added
 * Allow terra-date-picker to be a controlled component.
 * Added functionality to announce the pre-selected date to screen readers when navigating within the date picker.